hwcontext_vaapi: Improve logging around quirk detection
Clarify that the list is the naughty list, and therefore being on it is
not desirable.  The i965 driver does not need to be on the list after
version 2.0 (when the standard parameter buffer rendering behaviour was
changed).

(cherry picked from commit 852c7ba3f81f28b8b8b0930cf2cf2e10d7e9c4b0)