hwcontext_drm: Add AVDRMFrameDescriptor.format field
A AVDRMFrameDescriptor for a NV12 frame may be described in
a single layer descriptor with multiple planes,

(AVDRMFrameDescriptor) {
    .nb_layers = 1,
    .layers[0] = {
        .format           = DRM_FORMAT_NV12,
        .nb_planes        = 2,
        .planes[0] = {
            .object_index = 0,
        },
        .planes[1] = {
            .object_index = 0,
        },
    },
}

or a multi-layer descriptor with one plane in each layer.

(AVDRMFrameDescriptor) {
    .nb_layers = 2,
    .layers[0] = {
        .format           = DRM_FORMAT_R8,
        .nb_planes        = 1,
        .planes[0] = {
            .object_index = 0,
        },
    },
    .layers[1] = {
        .format           = DRM_FORMAT_RG88,
        .nb_planes        = 1,
        .planes[0] = {
            .object_index = 1,
        },
    },
}

With a multi-layer descriptor, the frame format is missing.

Add a AVDRMFrameDescriptor.format field to remove any ambiguity of
what frame format a multi-layer descriptor may have.

Signed-off-by: Jonas Karlman <jonas@kwiboo.se>