avcodec/v4l2m2m: clean up buffer options and pick sane defaults
Previously the default values for output/capture buffers were
quite high, causing a lot of memory usage on devices with limited
resources. On a default RPI3 installation, it is easy to run out of
kernel CMA memory without additional kernel boot flags.

Similar patches are being used in the LibreELEC ffmpeg fork for
AMLogic support.

I also reduced the minimum value for the buffering options so that users
can further reduce the number of buffers if required on their specific
platform.

Prior to this change, setting up a decode+scale+encode pipeline on
the RPI would fail with ENOMEM due to a total of 76 buffers being
pre-allocated.

Signed-off-by: Aman Gupta <aman@tmm1.net>